[FIX] Minor change to allow AMP training

+++ fairseq/models/wav2vec/wav2vec2.py
@@ -455,7 +455,7 @@ class Wav2Vec2Model(BaseFairseqModel):
                     mask_dropout=self.cfg.mask_dropout,
                 )
                 mask_indices = torch.from_numpy(mask_indices).to(x.device)
-            x = index_put(x, mask_indices, self.mask_emb)
+            x = index_put(x, mask_indices, self.mask_emb.to(x.dtype))
         else:
             mask_indices = None
